Feat: vehicle management from UI
