+++ src/pages/vehicle/catalog.html
@@ -1,7 +1,7 @@
 <p>
     Vehicle catalogue
 </p>
-<button onclick="navigate('newvehicle')">Add car</button>
+<button onclick ="location.href= './vehicle/?id=new';">Add car</button>
 <table>
     <tr>
       <th>Image</th>
