--- src/pages/vehicle/newvehicle.html
@@ -1,44 +1,0 @@
-<div>
-    <h2>Vehicle information</h2>
-    <p>ID: (unassigned)</p>
-    <form id="newVehicleForm">
-        <label for="make">Make</label><br>
-        <input type="text" id="make" name="make" placeholder="Vehicle make"><br><br>
-
-        <label for="model">Model</label><br>
-        <input type="text" id="model" name="model" placeholder="Vehicle model"><br><br>
-
-        <label for="type">Car type</label><br>
-        <select name="type" id="type">
-            <option value="Compact">Compact</option>
-            <option value="Estate">Estate</option>
-            <option value="Sedan">Sedan</option>
-            <option value="SUV">SUV</option>
-            <option value="Luxury">Luxury</option>
-        </select><br><br>
-
-        <label for="fuelType">Fuel type</label><br>
-        <select name="fuelType" id="fuelType">
-            <option value="Petrol">Petrol</option>
-            <option value="Diesel">Diesel</option>
-            <option value="Electric">Electric</option>
-        </select><br><br>
-
-        <label for="gearboxType">Gearbox type</label><br>
-        <select name="gearboxType" id="gearboxType">
-            <option value="Manual">Manual</option>
-            <option value="Automatic">Automatic</option>
-        </select><br><br>
-
-        <label for="year">Vehicle year</label><br>
-        <input type="number" id="year" name="year" placeholder="Year"><br><br>
-
-        <label for="price">Vehicle price per day</label><br>
-        <input type="number" id="price" name="price" placeholder="Price"><br><br>
-
-        <label for="image">Images (separate by comma)</label><br>
-        <input type="text" id="image" name="image" placeholder="Vehicle image url"><br><br>
-        
-        <button type="submit">Submit</button><br>
-    </form>
-</div>
