+++ ui/admin/home/catalogue/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Administration</title>
+    <link rel="stylesheet" href="/admin/home/style.css">
+  </head>
+  <body>
+    <nav id="navigation">
+    </nav>
+    <div id="content">
+    </div>
+  </body>
+  <script src="/admin/home/app.js"></script>
+</html>
